chore: organize utils
