+++ src/config/util.ts
@@ -1,17 +1,5 @@
 const debug = require('debug')
 
-export function flatMap<T, U>(arr: T[], fn: (i: T) => U[]): U[] {
-  return arr.reduce((arr, i) => [...arr, ...fn(i)], [] as U[])
-}
-
-export function mapValues<T extends Record<string, any>, TResult>(obj: {[P in keyof T]: T[P]}, fn: (i: T[keyof T], k: keyof T) => TResult): {[P in keyof T]: TResult} {
-  return Object.entries(obj)
-  .reduce((o, [k, v]) => {
-    o[k] = fn(v as any, k as any)
-    return o
-  }, {} as any)
-}
-
 export function resolvePackage(id: string, paths: { paths: string[] }): string {
   return require.resolve(id, paths)
 }
