+++ src/help/index.ts
@@ -5,7 +5,7 @@ import {Command} from '../command'
 import {CommandHelp} from './command'
 import {HelpFormatter} from './formatter'
 import RootHelp from './root'
-import {defaultFlagToCached} from '../default-flag-to-cached'
+import {defaultFlagToCached} from '../util/default-flag-to-cached'
 import {error} from '../errors'
 import {format} from 'node:util'
 import {load} from '../module-loader'
